pyproject.toml
[tool.poetry]
name = "lesson2"
version = "0.1.0"
description = ""
authors = ["dima cheban <dimacheban23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
uvicorn = "^0.32.1"
greenlet = "^3.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.10.2"}
python-multipart = "^0.0.19"
pathlib = {extras = ["context"], version = "^1.0.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
